Generic Typing to determine onClick, Downloads
