+++ screen2.0/src/app/downloads/downloadrange.tsx
@@ -1,13 +1,12 @@
-import { SetStateAction, useEffect, useState, useTransition } from "react"
+import { SetStateAction, useEffect, useState } from "react"
 import BiosampleTables from "../search/biosampletables"
-import { BIOSAMPLE_Data } from "../../common/lib/queries"
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2"
-import { Box, Button, Checkbox, CircularProgress, CircularProgressProps, FormControl, FormControlLabel, FormGroup, FormLabel, IconButton, MenuItem, Radio, RadioGroup, Stack, TextField, Typography } from "@mui/material"
+import { Box, Button, Checkbox, CircularProgress, CircularProgressProps, FormControl, FormControlLabel, FormGroup, FormLabel, IconButton, Radio, RadioGroup, Stack, TextField, Typography } from "@mui/material"
 import { downloadBED } from "../search/searchhelpers"
 import { parseGenomicRegion } from "../_mainsearch/parsegenomicregion"
 import { Close, Download } from "@mui/icons-material"
-import { ApolloQueryResult } from "@apollo/client"
-import { RegistryBiosample } from "../search/types"
+import GwasBiosampleTables from "../applets/gwas/gwasbiosampletables/gwasbiosampletables"
+import { RegistryBiosample } from "../applets/gwas/gwasbiosampletables/types"
 
 function CircularProgressWithLabel(
   props: CircularProgressProps & { value: number },
@@ -37,10 +36,6 @@ function CircularProgressWithLabel(
   );
 }
 
-interface DownloadRangeProps {
-  biosampleData: ApolloQueryResult<BIOSAMPLE_Data>
-}
-
 /**
  * @todo add linked genes info to query
  * @todo add genome switch to query mouse data. CONSERVATION data not available in mouse
@@ -48,12 +43,14 @@ interface DownloadRangeProps {
  * @todo make query only request the needed data
  * @todo allow changing the distance of linked genes searched
  */
-export const DownloadRange: React.FC<DownloadRangeProps> = ({biosampleData}) => {
+export const DownloadRange: React.FC = () => {
   const [assembly, setAssembly] = useState<"GRCh38" | "mm10">("GRCh38")
   const [inputValue, setInputValue] = useState<string>('chr11:5205263-5381894')
-  //Only reason this is an array is to easily interface with BiosampleTables
-  const [selectedBiosample, setSelectedBiosample] = useState<RegistryBiosample[]>([])
+  const [selectedBiosample, setSelectedBiosample] = useState<RegistryBiosample>(null)
   const [bedLoadingPercent, setBedLoadingPercent] = useState<number>(null)
+  //Used to disable assay checkboxes
+  const [availableAssays, setAvailableAssays] = useState<{ atac: boolean, ctcf: boolean, dnase: boolean, h3k27ac: boolean, h3k4me3: boolean }>({ atac: true, ctcf: true, dnase: true, h3k27ac: true, h3k4me3: true })
+  //Checkboxes
   const [selectedAssays, setSelectedAssays] = useState<{ atac: boolean, ctcf: boolean, dnase: boolean, h3k27ac: boolean, h3k4me3: boolean }>({ atac: true, ctcf: true, dnase: true, h3k27ac: true, h3k4me3: true })
   const [selectedConservation, setSelectedConservation] = useState<{ primate: boolean, mammal: boolean, vertebrate: boolean }>({ primate: true, mammal: true, vertebrate: true })
   // const [linkedGenes, setLinkedGenes] = useState<{ distancePC: boolean, distanceAll: boolean, ctcfChiaPet: boolean, rnapiiChiaPet: boolean }>({ distancePC: true, distanceAll: true, ctcfChiaPet: true, rnapiiChiaPet: true })
@@ -70,35 +67,50 @@ export const DownloadRange: React.FC<DownloadRangeProps> = ({biosampleData}) => 
       setSelectedConservation({ primate: true, mammal: true, vertebrate: true })
       // setLinkedGenes({ distancePC: true, distanceAll: true, ctcfChiaPet: true, rnapiiChiaPet: true })
     }
-    setSelectedBiosample([]);
+    setSelectedBiosample(null);
     (value === "GRCh38" || value === "mm10") && setAssembly(value)
   }
 
-  useEffect(() => {
-    if (selectedBiosample.length > 0) {
+  const handleSetSelectedBiosample = (biosample: RegistryBiosample) => {
+    setSelectedBiosample(biosample)
+    if (biosample) {
+      console.log(biosample)
       setSelectedAssays(
         {
-          atac: selectedBiosample[0].atac ? !!selectedBiosample[0].atac : null,
-          ctcf: selectedBiosample[0].ctcf ? !!selectedBiosample[0].ctcf : null,
-          dnase: selectedBiosample[0].dnase ? !!selectedBiosample[0].dnase : null,
-          h3k27ac: selectedBiosample[0].h3k27ac ? !!selectedBiosample[0].h3k27ac : null,
-          h3k4me3: selectedBiosample[0].h3k4me3 ? !!selectedBiosample[0].h3k4me3 : null
+          atac: !!biosample.atac,
+          ctcf: !!biosample.ctcf,
+          dnase: !!biosample.dnase,
+          h3k27ac: !!biosample.h3k27ac,
+          h3k4me3: !!biosample.h3k4me3
+        }
+      )
+      setAvailableAssays(
+        {
+          atac: biosample.atac !== null,
+          ctcf: biosample.ctcf !== null,
+          dnase: biosample.dnase !== null,
+          h3k27ac: biosample.h3k27ac !== null,
+          h3k4me3: biosample.h3k4me3 !== null
         }
       )
     } else {
       setSelectedAssays({ atac: true, ctcf: true, dnase: true, h3k27ac: true, h3k4me3: true })
+      setAvailableAssays({ atac: true, ctcf: true, dnase: true, h3k27ac: true, h3k4me3: true })
     }
-  }, [selectedBiosample])
+  }
 
   const handleDownloadBed = () => {
     const region = parseGenomicRegion(inputValue)
 
+    /**
+     * @todo make sure this works
+     */
     downloadBED(
       assembly,
       region.chromosome,
       +region.start,
       +region.end,
-      selectedBiosample.length > 0 && selectedBiosample[0],
+      selectedBiosample?.name,
       false,
       null,
       selectedAssays,
@@ -110,14 +122,10 @@ export const DownloadRange: React.FC<DownloadRangeProps> = ({biosampleData}) => 
   return (
     <Grid2 container spacing={3}>
       <Grid2 xs={6}>
-        <BiosampleTables
-          showRNAseq={false}
-          showDownloads={false}
-          biosampleSelectMode="replace"
-          biosampleData={biosampleData}
+        <GwasBiosampleTables
           assembly={assembly}
-          selectedBiosamples={selectedBiosample}
-          setSelectedBiosamples={setSelectedBiosample}
+          selected={selectedBiosample?.name}
+          onBiosampleClicked={(selected: RegistryBiosample) => handleSetSelectedBiosample(selected)}
         />
       </Grid2>
       <Grid2 xs={6}>
@@ -142,9 +150,9 @@ export const DownloadRange: React.FC<DownloadRangeProps> = ({biosampleData}) => 
           onChange={handleChange}
         />
         <Stack mt={1} direction="row" alignItems={"center"}>
-          <Typography>{`Selected Biosample: ${selectedBiosample.length > 0 ? selectedBiosample[0].displayname : "none"}`}</Typography>
-          {selectedBiosample.length > 0 &&
-            <IconButton onClick={() => setSelectedBiosample([])}>
+          <Typography>{`Selected Biosample: ${selectedBiosample ? selectedBiosample.displayname : "none"}`}</Typography>
+          {selectedBiosample &&
+            <IconButton onClick={() => handleSetSelectedBiosample(null)}>
               <Close />
             </IconButton>
           }
@@ -155,22 +163,20 @@ export const DownloadRange: React.FC<DownloadRangeProps> = ({biosampleData}) => 
               label="Assays"
               control={
                 <Checkbox
-                  checked={Object.values(selectedAssays).every(assay => assay || assay === null)}
+                  //For every available assay, check to make sure it's selected
+                  checked={Object.entries(availableAssays).every((assay: [string, boolean]) => !assay[1] || selectedAssays[assay[0]])} 
                   indeterminate={
-                    !(
-                      //If every value is not true/null
-                      Object.values(selectedAssays).every(assay => assay || assay === null) ||
-                      //Or every value is not false/null
-                      Object.values(selectedAssays).every(assay => !assay || assay === null)
-                    )
+                    //True if all available assays are not either true/false at the same time
+                    !Object.entries(availableAssays).every((assay: [string, boolean]) => !assay[1] || selectedAssays[assay[0]])
+                    && !Object.entries(availableAssays).every((assay: [string, boolean]) => !assay[1] || !selectedAssays[assay[0]])
                   }
                   onChange={(_, checked: boolean) => setSelectedAssays(
                     {
-                      atac: selectedAssays.atac === null ? null : checked,
-                      ctcf: selectedAssays.ctcf === null ? null : checked,
-                      dnase: selectedAssays.dnase === null ? null : checked,
-                      h3k27ac: selectedAssays.h3k27ac === null ? null : checked,
-                      h3k4me3: selectedAssays.h3k4me3 === null ? null : checked
+                      atac: availableAssays.atac ? checked : false,
+                      ctcf: availableAssays.ctcf ? checked : false,
+                      dnase: availableAssays.dnase ? checked : false,
+                      h3k27ac: availableAssays.h3k27ac ? checked : false,
+                      h3k4me3: availableAssays.h3k4me3 ? checked : false
                     }
                   )}
                 />
@@ -181,35 +187,35 @@ export const DownloadRange: React.FC<DownloadRangeProps> = ({biosampleData}) => 
                 onChange={(_, checked: boolean) => setSelectedAssays({ ...selectedAssays, dnase: checked })}
                 control={<Checkbox />}
                 label={"DNase"}
-                disabled={selectedAssays.dnase === null}
+                disabled={!availableAssays.dnase}
               />
               <FormControlLabel
                 checked={selectedAssays.atac}
                 onChange={(_, checked: boolean) => setSelectedAssays({ ...selectedAssays, atac: checked })}
                 control={<Checkbox />}
                 label={"ATAC"}
-                disabled={selectedAssays.atac === null}
+                disabled={!availableAssays.atac}
               />
               <FormControlLabel
                 checked={selectedAssays.ctcf}
                 onChange={(_, checked: boolean) => setSelectedAssays({ ...selectedAssays, ctcf: checked })}
                 control={<Checkbox />}
                 label={"CTCF"}
-                disabled={selectedAssays.ctcf === null}
+                disabled={!availableAssays.ctcf}
               />
               <FormControlLabel
                 checked={selectedAssays.h3k27ac}
                 onChange={(_, checked: boolean) => setSelectedAssays({ ...selectedAssays, h3k27ac: checked })}
                 control={<Checkbox />}
                 label={"H3K27ac"}
-                disabled={selectedAssays.h3k27ac === null}
+                disabled={!availableAssays.h3k27ac}
               />
               <FormControlLabel
                 checked={selectedAssays.h3k4me3}
                 onChange={(_, checked: boolean) => setSelectedAssays({ ...selectedAssays, h3k4me3: checked })}
                 control={<Checkbox />}
                 label={"H3K4me3"}
-                disabled={selectedAssays.h3k4me3 === null}
+                disabled={!availableAssays.h3k4me3}
               />
             </Box>
           </div>
